main: u16 for paging

+++ src/common/src/page.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Page {
     pub from: u16,
-    pub size: u8,
+    pub size: u16,
 }
